# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Librarium API",
        "version": "1.0.0",
        "description": "Library management API. Errors share one envelope: {\"error\": {\"code\", \"message\", \"details\", \"trace_id\"}}. A 422 means the input is malformed, a 409 means the input is fine and the business state disagrees. Every response carries an X-Request-Id header, which echoes an inbound one."
    },
    "servers": [
        {
            "url": "http://localhost/api/v1"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/authors/{author}": {
            "get": {
                "operationId": "authors.show",
                "summary": "Show an author",
                "tags": [
                    "Author"
                ],
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`AuthorResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AuthorResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "An error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Error overview.",
                                            "examples": [
                                                ""
                                            ]
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "put": {
                "operationId": "authors.update",
                "summary": "Update an author",
                "tags": [
                    "Author"
                ],
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "required": true,
                        "description": "The author ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAuthorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AuthorResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AuthorResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "authors.destroy",
                "summary": "Delete an author",
                "tags": [
                    "Author"
                ],
                "parameters": [
                    {
                        "name": "author",
                        "in": "path",
                        "required": true,
                        "description": "The author ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "409": {
                        "description": "The request is well formed, but a business rule refuses it.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "description": "Stable machine-readable code. One of: AUTHOR_IN_USE.",
                                                    "enum": [
                                                        "AUTHOR_IN_USE"
                                                    ]
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "description": "Human-readable message, safe to display."
                                                },
                                                "details": {
                                                    "type": "object",
                                                    "description": "Machine-readable context, when the error carries any."
                                                },
                                                "trace_id": {
                                                    "type": "string",
                                                    "description": "Correlates this response with the server logs; echoed as X-Request-Id."
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "message",
                                                "trace_id"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/books/{book}": {
            "get": {
                "operationId": "books.show",
                "description": "Carries an ETag; a conditional request with If-None-Match answers 304.",
                "summary": "Show a book",
                "tags": [
                    "Book"
                ],
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`BookResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BookResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "An error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Error overview.",
                                            "examples": [
                                                ""
                                            ]
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "put": {
                "operationId": "books.update",
                "description": "`available_copies` is system-managed and not accepted here; changing\n`total_copies` moves the counter by the same delta.",
                "summary": "Replace a book",
                "tags": [
                    "Book"
                ],
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "required": true,
                        "description": "The book ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`BookResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BookResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "description": "The stock cannot be cut below the copies currently on loan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "description": "Stable machine-readable code. One of: BOOK_COPIES_BELOW_LOANED.",
                                                    "enum": [
                                                        "BOOK_COPIES_BELOW_LOANED"
                                                    ]
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "description": "Human-readable message, safe to display."
                                                },
                                                "details": {
                                                    "type": "object",
                                                    "description": "Machine-readable context, when the error carries any."
                                                },
                                                "trace_id": {
                                                    "type": "string",
                                                    "description": "Correlates this response with the server logs; echoed as X-Request-Id."
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "message",
                                                "trace_id"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "books.destroy",
                "description": "Soft delete, so closed loans keep a row to point at.",
                "summary": "Delete a book",
                "tags": [
                    "Book"
                ],
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "required": true,
                        "description": "The book ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "409": {
                        "description": "The request is well formed, but a business rule refuses it.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "description": "Stable machine-readable code. One of: BOOK_HAS_ACTIVE_LOANS.",
                                                    "enum": [
                                                        "BOOK_HAS_ACTIVE_LOANS"
                                                    ]
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "description": "Human-readable message, safe to display."
                                                },
                                                "details": {
                                                    "type": "object",
                                                    "description": "Machine-readable context, when the error carries any."
                                                },
                                                "trace_id": {
                                                    "type": "string",
                                                    "description": "Correlates this response with the server logs; echoed as X-Request-Id."
                                                }
                                            },
                                            "required": [
                                                "code",
                                                "message",
                                                "trace_id"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "AuthorResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "bio": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "birth_year": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "books_count": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "name",
                    "bio",
                    "birth_year",
                    "created_at",
                    "updated_at"
                ],
                "title": "AuthorResource"
            },
...
            "StoreAuthorRequest": {
                "type": "object",
                "description": "FR-AUTHOR-1: name uniqueness is deliberately not enforced \u2014 homonymous\nauthors exist and rejecting them would be wrong.",
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "bio": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 2000
                    },
                    "birth_year": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "minimum": 0,
                        "maximum": 2026
                    }
                },
                "required": [
                    "name"
                ],
                "title": "StoreAuthorRequest"
            },
...
        },
        "responses": {
...
            "AuthenticationException": {
                "description": "No valid bearer token was supplied.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "description": "Stable machine-readable code. One of: UNAUTHENTICATED.",
                                            "enum": [
                                                "UNAUTHENTICATED"
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable message, safe to display."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Machine-readable context, when the error carries any."
                                        },
                                        "trace_id": {
                                            "type": "string",
                                            "description": "Correlates this response with the server logs; echoed as X-Request-Id."
                                        }
                                    },
                                    "required": [
                                        "code",
                                        "message",
                                        "trace_id"
                                    ]
                                }
                            },
                            "required": [
                                "error"
                            ]
                        }
                    }
                }
            },
            "AuthorizationException": {
                "description": "The authenticated role may not perform this action.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "description": "Stable machine-readable code. One of: FORBIDDEN.",
                                            "enum": [
                                                "FORBIDDEN"
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable message, safe to display."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Machine-readable context, when the error carries any."
                                        },
                                        "trace_id": {
                                            "type": "string",
                                            "description": "Correlates this response with the server logs; echoed as X-Request-Id."
                                        }
                                    },
                                    "required": [
                                        "code",
                                        "message",
                                        "trace_id"
                                    ]
                                }
                            },
                            "required": [
                                "error"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "The resource does not exist, or the caller may not know that it does.",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "type": "string",
                                            "description": "Stable machine-readable code. One of: NOT_FOUND.",
                                            "enum": [
                                                "NOT_FOUND"
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "description": "Human-readable message, safe to display."
                                        },
                                        "details": {
                                            "type": "object",
                                            "description": "Machine-readable context, when the error carries any."
                                        },
                                        "trace_id": {
                                            "type": "string",
                                            "description": "Correlates this response with the server logs; echoed as X-Request-Id."
                                        }
                                    },
                                    "required": [
                                        "code",
                                        "message",
                                        "trace_id"
                                    ]
                                }
                            },
                            "required": [
                                "error"
                            ]
                        }
                    }
                }
            }
        }
    }
}